[Core][View][Dissasm] # fixed warnings

+++ GViewCore/src/View/DissasmViewer/DissasmX86.cpp
@@ -37,8 +37,7 @@ bool Instance::DrawDissasmZone(DrawLineInfo& dli, DissasmCodeZone* zone)
 
     if (dli.textLineToDraw == 0)
     {
-        dli.renderer.WriteSingleLineText(
-              Layout.startingTextLineOffset, dli.screenLineToDraw + 1, "Dissasm zone", config.Colors.StructureColor);
+        dli.renderer.WriteSingleLineText(Layout.startingTextLineOffset, dli.screenLineToDraw + 1, "Dissasm zone", config.Colors.StructureColor);
         RegisterStructureCollapseButton(dli, zone->isCollapsed ? SpecialChars::TriangleRight : SpecialChars::TriangleLeft, zone);
         AdjustZoneExtendedSize(zone, 100);
         return true;
@@ -95,13 +94,12 @@ bool Instance::DrawDissasmZone(DrawLineInfo& dli, DissasmCodeZone* zone)
         return false;
     }
 
-    const auto instructionData = obj->GetData().Get(latestOffset, remainingZoneSize, false);
+    const auto instructionData = obj->GetData().Get(latestOffset, static_cast<uint32>(remainingZoneSize), false);
     if (!instructionData.IsValid())
         return true;
 
     cs_insn* insn;
-    size_t count = cs_disasm(
-          handle, instructionData.GetData(), instructionData.GetLength(), zone->zoneDetails.entryPoint, DISSASM_MAX_CACHED_LINES, &insn);
+    size_t count = cs_disasm(handle, instructionData.GetData(), instructionData.GetLength(), zone->zoneDetails.entryPoint, DISSASM_MAX_CACHED_LINES, &insn);
     if (count > 0)
     {
         LocalString<192> string;
